% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cubic_terms.R
\name{add_cubic_terms}
\alias{add_cubic_terms}
\title{Creates cubic spline functional forms for all predictors}
\usage{
add_cubic_terms(X)
}
\arguments{
\item{X}{design matrix of continuous predictors}
}
\value{
design matrix where each predictor has been modeled using a degree
three B-spline basis with the bs() function
}
\description{
Assumes all predictors for given design matrix are continuous, and
creates a matrix where every predictor has been modeled using a degree three
B-spline basis with the bs() function
}
\examples{
add_cubic_terms(X)

}
