% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quad_terms.R
\name{add_quad_terms}
\alias{add_quad_terms}
\title{Creates quadratic functional forms for all predictors}
\usage{
add_quad_terms(X)
}
\arguments{
\item{X}{design matrix of continuous predictors}
}
\value{
design matrix where each predictor has been modeled using a quadratic
term that does not include a linear component for each predictor
}
\description{
Assumes all predictors for given design matrix are continuous, and
creates a matrix where every predictor has been modeled using a quadratic
term that does not include a linear term for each predictor
}
\examples{
add_quad_terms(X)

}
