% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quart_terms.R
\name{add_quart_terms}
\alias{add_quart_terms}
\title{Creates quartile functional forms for all predictors}
\usage{
add_quart_terms(X)
}
\arguments{
\item{X}{design matrix of continuous predictors}
}
\value{
design matrix where each predictor has been modeled using
quartiles
}
\description{
Assumes all predictors for given design matrix are continuous, and
creates a matrix where every predictor has been modeled using quartiles
}
\examples{
add_quart_terms(X)

}
