% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{atoms}
\alias{atoms}
\alias{atoms,ripr::continuous_dist-method}
\title{Atoms of a distribution}
\usage{
atoms(x)
}
\arguments{
\item{x}{A \link{distribution} over a parameter space.}
}
\value{
\verb{(C, d)} numeric matrix, one atom per row.
}
\description{
Atoms of a distribution
}
\examples{
atoms(finite_dist(atoms = rbind(c(0.6, 0.4), c(0.2, 0.8)), weights = c(0.5, 0.5)))
}
