% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_RIPR_mat.R
\name{build_RIPR_mat}
\alias{build_RIPR_mat}
\title{Builds matrix for RIPR algorithm}
\usage{
build_RIPR_mat(X)
}
\arguments{
\item{X}{data frame of predictors, which can include both continous and
categorical predictors. Categorical predictors must be of class factor
and continuous predictors must be of class numeric}
}
\value{
updated design matrix where each continuous predictor has been
modeled using linear, quadratic, quartile, and cubic functional forms.
All categorical predictors appear at the end of this updated design matrix,
after all functional forms of the continuous predictors
}
\description{
Takes design matrix and creates updated design matrix where each continuous
predictor has been modeled with linear, quadratic, quartile, and cubic
functional forms
}
\examples{
build_RIPR_mat(X)

}
