% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{em_objective}
\alias{em_objective}
\title{EM M-step objective \verb{E_Q[r_k(X) log p_theta(X)]} as a face-oracle spec}
\usage{
em_objective(engine, family, log_r_k)
}
\arguments{
\item{engine}{An \code{expectation_engine}.}

\item{family}{A \code{family}.}

\item{log_r_k}{Length-M log-responsibility column of atom k.}
}
\value{
List of oracle objective closures.
}
\description{
As \code{\link[=fw_objective]{fw_objective()}}, for the responsibility-weighted M-step: \code{log_r_k} is
atom k's log-responsibility column.
}
\keyword{internal}
