% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_id}
\alias{engine_id}
\title{Stable identifier for the engine}
\usage{
engine_id(engine)
}
\arguments{
\item{engine}{An \code{expectation_engine}.}
}
\value{
Character scalar.
}
\description{
Stable identifier for the engine
}
\keyword{internal}
