% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{face_jacobian}
\alias{face_jacobian}
\title{Constant Jacobian of \code{\link[=parametrise]{parametrise()}}}
\usage{
face_jacobian(face)
}
\arguments{
\item{face}{A \code{face}.}
}
\value{
\verb{(d, V)} Jacobian matrix.
}
\description{
Constant Jacobian of \code{\link[=parametrise]{parametrise()}}
}
\keyword{internal}
