% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{flat_subnull}
\alias{flat_subnull}
\title{Which subnull each column of \code{\link[=flat_atoms]{flat_atoms()}} belongs to}
\usage{
flat_subnull(state)
}
\arguments{
\item{state}{A \link{ripr_state}.}
}
\value{
Integer vector.
}
\description{
Which subnull each column of \code{\link[=flat_atoms]{flat_atoms()}} belongs to
}
