% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{fw_line_search}
\alias{fw_line_search}
\title{Optimal mixing weight eps in \verb{[1e-10, 1-1e-10]} for adding a new atom to P_w}
\usage{
fw_line_search(engine, log_Pw, log_tm_new, tol = 1e-12)
}
\arguments{
\item{engine}{An \code{expectation_engine}.}

\item{log_Pw}{Length-M mixture log-density.}

\item{log_tm_new}{Length-M log-density of the new atom.}

\item{tol}{Line-search tolerance.}
}
\value{
Numeric mixing weight.
}
\description{
Optimal mixing weight eps in \verb{[1e-10, 1-1e-10]} for adding a new atom to P_w
}
\keyword{internal}
