% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{induced_draw}
\alias{induced_draw}
\alias{induced_draw,ripr::distribution,ripr::parametric_family-method}
\title{Draw from the induced mixture}
\usage{
induced_draw(mixing, family, n_obs)
}
\arguments{
\item{mixing}{A \link{distribution} over the parameter space.}

\item{family}{A \link{parametric_family}.}

\item{n_obs}{Number of draws.}
}
\value{
\verb{(n_obs, K)} numeric matrix.
}
\description{
Sampling from a mixture can be done easily by sampling once
from the mixing distribution, then sampling from the family at that parameter
value.
}
\keyword{internal}
