% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{init_point}
\alias{init_point}
\title{A starting atom on the subnull}
\usage{
init_point(subnull, ref)
}
\arguments{
\item{subnull}{A \link{subnull}.}

\item{ref}{Reference parameter vector.}
}
\value{
A parameter vector in the subnull.
}
\description{
Defaults to projecting a reference point, typically the alternative's mean.
}
