% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{is_finite_space}
\alias{is_finite_space}
\title{Is this sample space finite?}
\usage{
is_finite_space(space)
}
\arguments{
\item{space}{A \link{space}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Whether a sample space may be enumerated.
}
\examples{
is_finite_space(count_space(n_trials = 4L, k = 3L))
is_finite_space(real_region(1L))
}
\seealso{
\code{\link[=enumerate_space]{enumerate_space()}}
}
