% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{is_finite_support}
\alias{is_finite_support}
\title{Does this family have a finite, enumerable sample space?}
\usage{
is_finite_support(family)
}
\arguments{
\item{family}{A \link{sampling_family}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does this family have a finite, enumerable sample space?
}
