% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinomial_family}
\alias{multinomial_family}
\title{Multinomial sampling family}
\usage{
multinomial_family(n_trials, k)
}
\arguments{
\item{n_trials}{Integer. Trials per observation.}

\item{k}{Integer. Number of categories.}
}
\value{
A \code{multinomial_family}.
}
\description{
The \code{n_trials}-trial, \code{k}-category multinomial over \code{\link[=count_space]{count_space()}}, with
parameter space equal to the full probability simplex.
}
\examples{
multinomial_family(n_trials = 20L, k = 3L)
}
