% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{n_atoms}
\alias{n_atoms}
\alias{n_atoms,ripr::continuous_dist-method}
\title{Number of atoms in a distribution}
\usage{
n_atoms(x)
}
\arguments{
\item{x}{A \link{distribution} over a parameter space.}
}
\value{
Integer.
}
\description{
\code{NA} for a continuous measure.
}
\examples{
n_atoms(dirac(theta = c(0.5, 0.5)))
n_atoms(finite_dist(atoms = rbind(c(0.6, 0.4), c(0.2, 0.8)), weights = c(0.5, 0.5)))
}
