% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{n_cells}
\alias{n_cells}
\title{Number of convex regions a region decomposes into}
\usage{
n_cells(space)
}
\arguments{
\item{space}{A \link{region}.}
}
\value{
Integer.
}
\description{
Number of convex regions a region decomposes into
}
\examples{
n_cells(simplex_region(vertices = diag(3)))
}
