% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{n_outcomes}
\alias{n_outcomes}
\title{Number of outcomes in the engine's working set}
\usage{
n_outcomes(engine)
}
\arguments{
\item{engine}{An \code{expectation_engine}.}
}
\value{
Integer.
}
\description{
Number of outcomes in the engine's working set
}
\keyword{internal}
