% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{n_parts}
\alias{n_parts}
\title{Number of convex regions a region was declared as}
\usage{
n_parts(space)
}
\arguments{
\item{space}{A \link{region}.}
}
\value{
Integer.
}
\description{
Number of convex regions a region was declared as
}
\examples{
n_parts(simplex_region(vertices = diag(3)))
}
