% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{param_dim}
\alias{param_dim}
\title{Dimension of the parameter vector}
\usage{
param_dim(family)
}
\arguments{
\item{family}{A \code{family}.}
}
\value{
Integer parameter dimension.
}
\description{
Dimension of the parameter vector
}
