% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative.R
\name{point_alt}
\alias{point_alt}
\title{Point alternative \verb{Q = p_\{theta*\}}}
\usage{
point_alt(theta_star = integer(0))
}
\arguments{
\item{theta_star}{Numeric parameter vector of the single alternative.}
}
\value{
A \code{point_alt}.
}
\description{
Point alternative \verb{Q = p_\{theta*\}}
}
