% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{q_positive_mask}
\alias{q_positive_mask}
\title{Mask of outcomes with positive Q-mass}
\usage{
q_positive_mask(engine)
}
\arguments{
\item{engine}{An \code{expectation_engine}.}
}
\value{
Logical vector of length M.
}
\description{
Mask of outcomes with positive Q-mass
}
\keyword{internal}
