% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_space.R
\name{real_space}
\alias{real_space}
\title{The space of \code{d}-dimensional real numbers: \eqn{\mathbb{R}^d}{R^d}}
\usage{
real_space(d)
}
\arguments{
\item{d}{Integer dimension.}
}
\value{
A \code{real_space}.
}
\description{
The space of \code{d}-dimensional real numbers: \eqn{\mathbb{R}^d}{R^d}
}
\examples{
real_space(2L)
}
