% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIPR.R
\name{RIPR}
\alias{RIPR}
\title{Runs the RIPR algorithm}
\usage{
RIPR(X, Y)
}
\arguments{
\item{X}{data frame with all predictors of interest. Categorical and
continuous predictors can be included, where the class of the categorical
predictors must be factors, and the class of the continuous predictors
must be numeric.}

\item{Y}{response vector with length equal to the number of rows of X}
}
\value{
two-item list where the first item is the updated design matrix
used for the ridge regression fit of the RIPR algorithm, and the second item
is glmnet object for the ridge regression fit
}
\description{
Runs the RIPR algorithm/performs the appropriate ridge regression fit using
the updated design matrix including linear, quadratic, quartile, and cubic
functional forms for the continuous predictors
}
\examples{
RIPR(X,Y)

}
