% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_atoms}
\alias{state_atoms}
\title{Atom list (length-d parameter vectors)}
\usage{
state_atoms(state)
}
\arguments{
\item{state}{A \code{mixture_state}.}
}
\value{
List of parameter vectors.
}
\description{
Atom list (length-d parameter vectors)
}
\keyword{internal}
