% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_face_idx}
\alias{state_face_idx}
\title{Face index per atom}
\usage{
state_face_idx(state)
}
\arguments{
\item{state}{A \code{mixture_state}.}
}
\value{
Integer vector.
}
\description{
Face index per atom
}
\keyword{internal}
