% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_n_atoms}
\alias{state_n_atoms}
\title{Number of live atoms}
\usage{
state_n_atoms(state)
}
\arguments{
\item{state}{A \code{mixture_state}.}
}
\value{
Integer.
}
\description{
Number of live atoms
}
\keyword{internal}
