% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_weights}
\alias{state_weights}
\title{Mixture weights}
\usage{
state_weights(state)
}
\arguments{
\item{state}{A \code{mixture_state}.}
}
\value{
Numeric weight vector.
}
\description{
Mixture weights
}
\keyword{internal}
