% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_measure.R
\name{atoms}
\alias{atoms}
\title{Parameter atoms of a mixing measure}
\usage{
atoms(x)
}
\arguments{
\item{x}{A \link{mixing_measure}.}
}
\value{
\verb{(K, C)} numeric matrix.
}
\description{
Parameter atoms of a mixing measure
}
\examples{
atoms(finite_mixing(components = cbind(c(0.6, 0.4), c(0.2, 0.8)), weights = c(0.5, 0.5)))
}
