% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_measure.R
\name{atoms}
\alias{atoms}
\title{Parameter atoms of a mixing measure}
\usage{
atoms(x)
}
\arguments{
\item{x}{A \link{mixing_measure}.}
}
\value{
\verb{(K, C)} numeric matrix.
}
\description{
Parameter atoms of a mixing measure
}
