% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{contains}
\alias{contains}
\title{Does a parameter vector belong to the space?}
\usage{
contains(space, theta, tol = 1e-08)
}
\arguments{
\item{space}{A \link{convex_region}.}

\item{theta}{Parameter vector.}

\item{tol}{Tolerance.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does a parameter vector belong to the space?
}
\examples{
s <- simplex_region(vertices = diag(3))
contains(s, c(1 / 3, 1 / 3, 1 / 3))
contains(s, c(2, -1, 0))
}
