% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{contains}
\alias{contains}
\title{Is a parameter in the subnull?}
\usage{
contains(subnull, theta, tol = 1e-08)
}
\arguments{
\item{subnull}{A \link{subnull}.}

\item{theta}{Parameter vector.}

\item{tol}{Tolerance.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is a parameter in the subnull?
}
\examples{
s <- simplex_null(vertices = diag(3))
contains(s, c(1 / 3, 1 / 3, 1 / 3))
contains(s, c(2, -1, 0))
}
