% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{discretise}
\alias{discretise}
\title{Replace a distribution with a finite one comprised of atoms drawn from it}
\usage{
discretise(dist, n)
}
\arguments{
\item{dist}{A \link{distribution} to sample.}

\item{n}{Number of draws.}
}
\value{
A \link{finite_dist} on \code{n} equally weighted atoms.
}
\description{
Constructs a \link{finite_dist} comprising \code{n} draws from \code{dist} as equally
weighted atoms. It is the empirical distribution based on a sample drawn
from it. Converges to \code{dist} as \code{n} grows.
}
\details{
This can be used as a convenient way to approximate a mixture distribution
if no closed-form exists. See examples.
}
\examples{
set.seed(1)
fam <- multinomial_family(n_trials = 6L, k = 3L)

W <- dirichlet(alpha = c(4, 3, 2))
approx <- discretise(W, 500L)
n_atoms(approx)

x <- enumerate_space(fam@sample_space)
log_density(fam(W), c(2,2,2))
log_density(fam(approx), c(2,2,2))
}
