% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{dist_draw}
\alias{dist_draw}
\title{Draw outcomes from a distribution}
\usage{
dist_draw(dist, n_obs)
}
\arguments{
\item{dist}{An \link{outcome_distribution}.}

\item{n_obs}{Number of draws.}
}
\value{
\verb{(n_obs, K)} numeric matrix.
}
\description{
Draw outcomes from a distribution
}
