% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{flat_subnull}
\alias{flat_subnull}
\title{Which subnull each column of \code{\link[=flat_atoms]{flat_atoms()}} belongs to}
\usage{
flat_subnull(state)
}
\arguments{
\item{state}{A \link{ripr_state}.}
}
\value{
Integer vector.
}
\description{
Which subnull each column of \code{\link[=flat_atoms]{flat_atoms()}} belongs to
}
\examples{
fam <- multinomial_family(n_trials = 4L, k = 3L)
plurality <- null_model(
  fam,
  list(
    simplex_region(vertices = cbind(c(0.5, 0.5, 0), c(0, 1, 0), c(0, 0, 1))),
    simplex_region(vertices = cbind(c(0.5, 0, 0.5), c(0, 1, 0), c(0, 0, 1)))
  )
)
Q <- mixture(point_mixing(c(0.4, 0.35, 0.25)), fam)
state <- ripr_init(Q, plurality)
flat_subnull(state)
}
