% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{in_null}
\alias{in_null}
\title{Does any subnull contain this parameter?}
\usage{
in_null(null, theta, tol = 1e-08)
}
\arguments{
\item{null}{A \link{null_model}.}

\item{theta}{Parameter vector.}

\item{tol}{Tolerance.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does any subnull contain this parameter?
}
