% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{induced_draw}
\alias{induced_draw}
\title{Draw from the induced mixture}
\usage{
induced_draw(mixing, family, n_obs)
}
\arguments{
\item{mixing}{A \link{mixing_measure}.}

\item{family}{A \link{sampling_family}.}

\item{n_obs}{Number of draws.}
}
\value{
\verb{(n_obs, K)} numeric matrix.
}
\description{
Draw from the induced mixture
}
\keyword{internal}
