% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{induced_draw}
\alias{induced_draw}
\title{Draw outcomes from the induced marginal}
\usage{
induced_draw(mixing, family, n_obs)
}
\arguments{
\item{mixing}{A \link{mixing} measure.}

\item{family}{A \link{family}.}

\item{n_obs}{Number of draws.}
}
\value{
\verb{(n_obs, d)} numeric matrix of draws.
}
\description{
Draw outcomes from the induced marginal
}
\keyword{internal}
