% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{induced_log_density}
\alias{induced_log_density}
\title{Induced log density \verb{log int p_theta(x) dW(theta)}}
\usage{
induced_log_density(mixing, family, x)
}
\arguments{
\item{mixing}{A \link{mixing_measure}.}

\item{family}{A \link{sampling_family}.}

\item{x}{\verb{(M, K)} matrix of outcomes.}
}
\value{
Length-\code{M} numeric vector.
}
\description{
Induced log density \verb{log int p_theta(x) dW(theta)}
}
\keyword{internal}
