% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_space.R
\name{init_point}
\alias{init_point}
\title{A starting atom in a parameter space}
\usage{
init_point(space, ref)
}
\arguments{
\item{space}{A \link{parameter_space}.}

\item{ref}{Reference parameter vector.}
}
\value{
A parameter vector in the space.
}
\description{
Defaults to projecting a reference point, typically the alternative's mean.
}
\examples{
s <- simplex_region(vertices = diag(3))
init_point(s, c(1, 0, 0))
}
