% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_space.R
\name{is_finite_space}
\alias{is_finite_space}
\title{Is this sample space finite?}
\usage{
is_finite_space(space)
}
\arguments{
\item{space}{A \link{sample_space}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Use to determine whether a sample space may be enumerated.
}
\examples{
is_finite_space(count_space(n = 4L, k = 3L))
is_finite_space(real_space(1L))
}
\seealso{
\code{\link[=enumerate_space]{enumerate_space()}}
}
