% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_measure.R
\name{n_atoms}
\alias{n_atoms}
\title{Number of atoms in a mixing measure}
\usage{
n_atoms(x)
}
\arguments{
\item{x}{A \link{mixing_measure}.}
}
\value{
Integer.
}
\description{
Number of atoms in a mixing measure
}
