% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{n_nodes}
\alias{n_nodes}
\title{Number of quadrature nodes}
\usage{
n_nodes(engine)
}
\arguments{
\item{engine}{A \link{quadrature}.}
}
\value{
Integer.
}
\description{
Number of quadrature nodes
}
\examples{
fam <- multinomial_family(n_trials = 3L, k = 2L)
Q <- induced_distribution(fam, point_mixing(c(0.5, 0.5)))
n_nodes(resolve_engine(exact_engine(), Q, fam))
}
