% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.R
\name{n_subnulls}
\alias{n_subnulls}
\title{Number of convex pieces in a null}
\usage{
n_subnulls(null)
}
\arguments{
\item{null}{A \link{null_model}.}
}
\value{
Integer.
}
\description{
Number of convex pieces in a null
}
