% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{param_dim}
\alias{param_dim}
\title{Dimension of the parameter vector}
\usage{
param_dim(family)
}
\arguments{
\item{family}{A \link{sampling_family}.}
}
\value{
Integer parameter dimension.
}
\description{
Dimension of the parameter vector
}
\examples{
param_dim(multinomial_family(n_trials = 4L, k = 3L))
}
