% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_space.R
\name{space_dim}
\alias{space_dim}
\title{The dimension of one element of the sample space}
\usage{
space_dim(space)
}
\arguments{
\item{space}{A \link{sample_space}.}
}
\value{
An integer.
}
\description{
The dimension of one element of the sample space
}
\examples{
space_dim(count_space(n = 4L, k = 3L))
space_dim(real_space(2L))
}
