% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_dim}
\alias{space_dim}
\title{The dimension of one element of a space}
\usage{
space_dim(space)
}
\arguments{
\item{space}{A \link{space}.}
}
\value{
An integer.
}
\description{
The dimension of one element of a space
}
\examples{
space_dim(count_space(n = 4L, k = 3L))
space_dim(simplex_region(vertices = diag(3)))
}
