% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{weights.distribution}
\alias{weights.distribution}
\alias{weights,ripr::continuous_dist-method}
\title{Weights of a distribution}
\arguments{
\item{object}{A \link{distribution} over a parameter space.}

\item{...}{Ignored.}
}
\value{
Numeric vector summing to 1.
}
\description{
Weights of a distribution
}
\examples{
weights(finite_dist(components = cbind(c(0.5, 0.5)), weights = 1))
}
