% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_measure.R
\name{weights.mixing_measure}
\alias{weights.mixing_measure}
\alias{weights,ripr::continuous_mixing-method}
\title{Weights of a mixing measure}
\arguments{
\item{object}{A \link{mixing_measure}.}

\item{...}{Ignored.}
}
\value{
Numeric vector summing to 1.
}
\description{
Weights of a mixing measure
}
\examples{
weights(finite_mixing(components = cbind(c(0.5, 0.5)), weights = 1))
}
